Fixed parabola collision detection bug.
TileEngine::calculateParabola uses straight line segments to approximate
collision detection of the parabola. But when a collection was detected,
instead of storing the coordinates of the collisions, it just stored the
coordinates of the end point of current line segment - which may or may
not have been where the collision occurred. And so sometimes the
end-point of the tragetory was inside the target, and sometimes it
wasn't - resulting in inconsistant behaviour.

calculateParabola now stores the coorinates of the actual point of
collision as the end point.