Fixed line-of-fire problems
TileEngine::canTargetUnit correctly determines whether or not a
character has line-of-fire to a target, but this information was not
being correctly used. The result was that the game often behaved as
though a character has line-of-fire when they actually didn't.

This is now fixed so that the return value of TileEngine::canTargetUnit
is the primary way of preventing shots with no line-of-fire.