Adjusted final voxel of throw trajectories to fix a problem.
Some maps are designed such that an object of height 23 is effectively
the floor of the tile above. This is a problem for thrown objects
because the object throw will automatically target the top tile, but the
collision will always be at the top of the bottom tile (at height 23,
whereas 24 is where the next tile starts). This results in a 'no arc of
throw' message.

To fix this, the final voxel of thrown objects is now increased by 1
(always); so that the thrown object lands "on top" of the collision
point rather than inside it.