Fixed a problem with the storage check when transferring items.
Previously, the storage space of both the sending and receiving base was
checked for all transfers. This meant that if the sending base was
already overfull, you could not send anything unless you sent enough to
stop it being overfull.

I've changed it to now only check the base whose available space is
decreasing, so that any transfer which gets us closer to not being
overfull is accepted.