Added config option to show/disable bar overflow layers.
The layered display for overflowing status bars (such as the stamina
bar) can now be turned on or off in the game options. The default is on,
with up to 2 overflow layers. (The colour of the second overflow layer
doesn't always suit the bar; but in practice its very rare to have more
than 1 overflow layer anyway.)

Option name is `Options::showBarOverflowLayers`
The text keys are `STR_SHOW_BAR_OVERFLOW` and `STR_SHOW_BAR_OVERFLOW_DESC`
(Both these text key are empty at this stage.)