Changed the draw order for duel-purpose bars with overflow
Previously we drew the bar for all of the first number, then drew all of
the second. This meant that overflow of the first number could be
covered by small amounts of the second.

Using stun & health as an example, the new draw order is this:
health up to 100, then stun up to 100, health to 200, stun to 200, etc.