Added item requirements text when placing base facilities.
When constructing base facilities, required items are now listed as part
of the building cost.

Up to three different items can be displayed. (If more than three items
types are required, only the first three will be shown. If the building
does not require any items, the text layout is unchanged from how it was
previously.)