Improved handling of unusual items in the tranfer screen.
Storage size is now counted and checked for live prisoners.
(Previously, only prison / alien containment was checked).

Storage limits are now checked when adding negative, so that the
player is warned if there is not enough storage when removing them.
(Previously this wasn't checked until the player pressed the confirm
 button.)