Fixed bug in TileEngine::voxelCheck related to stairs / slopes.
In cases where a unit was standing below the target tile, this
function was erroneously using the target tile's terrain height
instead of the terrain height of the unit's tile. The result is
that units standing on sloped terrain or stairs would often be
missed by this function.

The most obvious effect of this bug is that units standing on slopes
were sometimes impossible to hit with melee attacks. (Melee attacks
which were meant to register as a hit would deal no damage, because
voxelCheck would fail to find the target to deal the damage.)

The function now uses the correct terrain height; and so units on
slopes will probably be (correctly) hit more often than before!