Added option for overflow bars
A layered display for overflowing status bars (such as the stamina
bar) which can be turned on or off in the game options. The default is on,
with up to 2 overflow layers. (The colour of the second overflow layer
doesn't always suit the bar; but in practice its very rare to have more
than 1 overflow layer anyway.)