Changed tally system for killCriteria based commendations
Previously, the successes of each OR condition was counted separately and the
max value was used. In the new system, the successes of each OR condition are
added together.

For example:

killCriteria:
  -
     - [1, ["BT_GRENADE", "STATUS_DEAD"]]
  -
     - [1, ["BT_PROXIMITYGRENADE", "STATUS_DEAD"]]

Previously, the result would whichever was higher of kills with grenades, or
kills with proximity grenades. In the new system it is the sum of them; and
so we can award a commendation for total number of kills with grenades or
proximitygrenades.

For killsWithCriteriaCareer, all successes are added.
For killsWithCriteriaMission and killsWithCriteriaTurn, only one success is
possible for each mission/turn respectively.