Melee attacks from ranged weapons now identifiable in soldier stats
Kill stats record the weapon name and the ammo name used for each kill.
But until now, there was no way of telling if a gun's melee attack was used.
This led to incorrect damage types when awarding commendations.

To fix this, we now list 'BA_HIT' as the ammo type when the secondary melee
damage is used. Melee weapons and ammoless guns still put the weapon itself
as the ammo field so that their main damage data is used.

When checking commendation conditions, 'BA_HIT' indicates that the ranged
weapon's melee damage data should be used rather than the default data.