Changed behaviour of craft inventory screen for the new equipment system.
When using `oxceAlternateCraftEquipmentManagement`, entering the inventory
screen from a craft will now show all available equipment in the base. Changes
made the soldier's inventory are then used to update the items loaded on the
craft, so that the craft has all of the soldiers' items (as well as any
unequipped items already loaded onto the craft).

Previouly, entering the inventory screen from a craft would show only the
items that were already on-board the craft, and no changes made to soldiers
inventory would effect the items on the craft.

Note: these changes only apply when using the new equipment system; and also
only when using the craft inventory screen. In particular, changing soldiers'
equipment from the base inventory screen currently does not affect which items
are loaded onto any craft. (Although this is likely to be changed soon!)