Added a work-around for limited-space craft to see all base equipment.
The alternative equipment setting try to move all base equipment into the
craft temporarily for the equip screen. But if the craft couldn't hold all
the equipment, this was more confusing than helpful, because only some base
equipment was made available. This change basically just ignores the craft
space limit temporarily. (It's an easy but imperfect fix - because it allows
players to use the equipment screen to get a bit above the craft limit if
they really want.)